Fix job save; part of CONNECTORS-764.
git-svn-id: https://svn.apache.org/repos/asf/manifoldcf/trunk@1513789 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/pull-agent/src/main/java/org/apache/manifoldcf/crawler/jobs/Jobs.java
@@ -763,14 +763,14 @@ public class Jobs extends org.apache.manifoldcf.core.database.BaseTable
                 {
                   for (String linkType : (Collection<String>)filterRows.keySet())
                   {
-                    Integer oldCount = (Integer)filterRows.get(linkType);
-                    Integer newCount = (Integer)newFilterRows.get(linkType);
+                    Long oldCount = (Long)filterRows.get(linkType);
+                    Long newCount = (Long)newFilterRows.get(linkType);
                     if (oldCount == null || newCount == null)
                     {
                       isSame = false;
                       break;
                     }
-                    else if (oldCount.intValue() != newCount.intValue())
+                    else if (oldCount.longValue() != newCount.longValue())
                     {
                       isSame = false;
                       break;
